ROBUSTEZ! en, ver detalle.
ROBUSTEZ EN: devices listing, devices management, internet listing, employees management, internte management, movile telephony listing, myservices

+++ techsa_frontend/src/app/views/eployees-managment/eployees-managment.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AgenteVentasService } from '../../services/agente-ventas.service';
 import { GerenteService } from '../../services/gerente.service';
+import { AuthService } from 'src/app/services/auth.service';
 
 
 @Component({
@@ -14,24 +15,47 @@ export class EployeesManagmentComponent implements OnInit {
 
   constructor( 
     private agenteVentasService:AgenteVentasService,
-    private gerenteService: GerenteService
+    private gerenteService: GerenteService,
+    private auth:AuthService
     ) { }
 
   async ngOnInit() {
-    this.employees = await this.agenteVentasService.getAgenteVentasAll();
+    try {
+      this.employees = await this.agenteVentasService.getAgenteVentasAll();
+    } catch (err) {
+      alert("Error al cargar la lista de empleados.\n Intentelo de nuevo.");
+    }
     this.employees=this.employees.concat(await this.gerenteService.getGerenteAll());
     console.log(this.employees);
   }
 
   async delete (employeeID:string,employeePuesto:String){
     console.log("Eliminando empleado: ",employeeID,employeePuesto)
-    if (employeePuesto == 'Gerente'){
-      await this.gerenteService.deleteGerente(employeeID)
-    }else{
-      await this.agenteVentasService.deleteAgenteVentas(employeeID);
-    }
-    
-    
-  }
 
+    const token = localStorage.getItem('access_token');
+    const id = await this.auth.getUserId(token);
+    if( employeeID ==  id.slice(10,14)){
+      alert("Usted se encuentra logueado en este momento. \n No se puede eliminar su cuenta");
+    }else{
+      if( confirm("¿Desea eliminar esta cuenta?") ){
+        if (employeePuesto == 'Gerente'){
+          try {
+            await this.gerenteService.deleteGerente(employeeID)
+            alert("El gerente fue eliminado.");
+            location.reload();
+          } catch (err) {
+            alert("El gerente no pudo ser eliminado.\ Intente Nuevamente");
+          }
+        }else{
+          try {
+            await this.agenteVentasService.deleteAgenteVentas(employeeID);
+            alert("El Agente de Ventas fue eliminado.");
+            location.reload();
+          } catch (err) {
+            alert("El Agente de Ventas no pudo ser eliminado.\ Intente Nuevamente");
+          }
+        }
+      }
+    }
+  }
 }
